+++ src/Decorator/Domain/Drinks/HouseBlend.cs
@@ -7,6 +7,6 @@
             Description = "House Blend Coffee";
         }
 
-        public override double Cost() => 0.89;
+        public override decimal Cost() => 0.89M;
     }
 }
